Refactor parse_statements_job and tests
This commit refactors the parse_statements_job to remove the need
for the root_doc_dir environment variable. We now pass this value
in as an argument to .perform. It was difficult to cleanly test
this job when using an environment variable.

In the tests, we clean up mock statement parser class and extract
it into a new file. It was missing the required constants such as
BANK_NAME, ACCOUNT_NAME, and ACCOUNT_TYPE.

We also mock out the pdf reader more cleanly so that it doesn't
require a real PDF file, nor does it require us to overwrite the
@file_path anymore.