Backfill tests for Schwab Checking statement parser
This commit fully backfills tests for the Charles Schwab Checking
StatementParser class. There were two lines of logic that were
missed by test coverage: the multiple keywords detected as well as
no keywords detected in the transaction description. We add these
two tests here.

We also refactor a lof of the parser class to satisfy rubocop
offenses.

We also add the frozen string literal to the top of the test_helper
class to satisfy the rubocop offense.