Refactor categories
This refactors the category model to remove the check for destroy,
and add a dependent restrict with exception which will block
deletion if any transactions are associated with the category.

We update the categories controller to handle this exception and
reraise a bad request error which is the standard for our
controllers.

This also refactors the category_data_entity .get_data method to
.data.

Tests were backfilled for all edge cases and rubocop offenses were
fixed.