Refactor subcategories
This refactors the subcategory model to remove the check for
destroy, and add a dependent restrict_with_exception which
will block deletion if any transactions are associated with
the subcategory.

We update the subcategories controller to handle this exception
and reraise a bad request error which is the standard for our
controllers.

This also adds a new subcategory serializer to camelCase the
categoryId, and not return the updated_at and created_at fields.

Tests were backfilled for all edge cases and rubocop offenses were
fixed.