Refactor category data and definitions
This updates the category data entity to preload subcategories
along with categories from the database to avoid N+1 queries. We
also switch to using the Subcategory serailizer in the entity to
make the subcategory response consistent across all actions.

This also updates the frontend typescript definitions file to be
consistent with the backend. This commit also alphabetizes the
definition file types.