Rename FilterSelector and move type to Item
This renames the FilterSelector component to the correct name
rather than leaving it as UserSelector, which was the name of the
component before we generalized it to take in any type.

The FilterItemType is moved out of the definitions file and placed
into the FilterItem component since it's directly associated with
that component. The definitions file in lib should be reserved for
defintions related to backend items.