Major refactor of frontend files and api functions
This commit extracts the api functions in the frontend app into
individual lib/api files for each resource. Prior to this change,
all api functions to fetch and push data were in one "data" file,
but the frontend app is starting to grow so big with many different
resources that this is not sustainable.

We also move the shared components out of src/app and into a new
src/components directory. At this stage, the only components in the
/app directories are ones that are directly tied to the page that
they serve. All other shared components are in src/components.