Implement categorization-rule page filtering
This implements the filter functionality for the
categorization-rule page. There are two dropdown filters:
Subcategory and Account.

The search bar works by matching the search string on the following
fields:
description, amount, date
If the rule only uses account for the condition, it won't match at
this time.

The match logic has a scoring system and the results will be sorted
based on the strength of the match. Here is the list from highest
to lowest match:

1. Exact match (case-sensitive)
2. Substring match (case-sensitive)
3. Exact match (case-sensitive, ignoring non-word characters)
4. Substring match (case-sensitive, ignoring non-word characters)
5. Exact match (case-insensitive)
6. Substring match (case-insensitive)
7. Substring match (case-insensitive, ignoring non-word characters)
8. No match - will not appear in search results