Implement modify categorization rule logic
This implements all of the logic to modify a Categorization Rule.
You can add conditions, delete conditions, modify conditions,
and modify the subcategory.

If any of the conditions or subcategory is modified, then we'll go
with a rule update to use just one API call.

However, if no conditions or subcategory is modified, and
conditions are added or deleted, then we'll make individual API
calls for each added or deleted condition. This is to avoid
destroying and recreating all of the conditions every time on the
backend since that is how the update rule API works when conditions
are passed in.

The add new rule does not work yet and will be implemented in the
next commit.