Reset condition attributes and add validation
This commit changes the logic of the categorizationCondition
component in a way that sets the matchType and matchValue
attributes to empty if the transactionField is changed. This is
important because if the user didn't manually set a matchvalue when
the transactionField was set to account, the account would have an
invalid value that was not an account ID.

We also add validation to ensure that the user selects a non-empty
matchType and matchValue. If the fields are invalid, they will have
a red border around them to indicate so. If the user attepmts to
save the rule when these are invalid, the save will fail and an
error message will display in the lower right corner of the screen.

This commit also adds a new shared error component that we will
reuse throughout the entire site.

We also have a major bug fix for updating the rule. We need to
overwrite the rule conditions with the updated rules, new rules and
remove the deleted rules. Then we need to clear the state variables
that track modified conditions such as updatedConditions,
newConditions and deletedConditionIds.