Implement add conditions to empty rule
This implements the logic to handle adding a condition to an empty
rule from the view categorizationRuleRow component. All editing
functions to a rule must be done in the EditableCategorizationRule
componenent except for this exception, where we display an "Add
Condition" button to an empty rule. This places the rule into edit
mode and adds an empty condition.