Complete logic to add categorization rule
This completes the implementation of the logic to add a
categorization rule. We add a new prop called isNewRule which tells
the EditableCategorizationRule a few things:
1. It will call createCategorizationRule instead of updating an
existing rule.
2. We format the border to green and add a "new rule" badge.

We also refactor the update methods so that each one individually
handles the call to onSave rather than putting it all inside the
handleRuleSave function. This keeps the functions shorter and more
focused.

We also remove any logic that clears the updated state variables
after saving the rule because the parent will simply erase this
component once onSave is called.