Refactor subcategory param to always use ID
This refactors the transaction create and update endpoints to use
subcategoryId as a parameter rather than subcategory name.

With this change comes a refactor to the frontend components for
the category dropdown selector. We now pass the id and name as an
object so that we don't have to perform any lookups in the
components utilizing the dropdown. We are able to remove the helper
functions that convert id to name and vice versa.