Add category_type to category
This adds a new enum category_type to category. We will use this
field to distinguish between what is an income category versus
spend or transfer. These will be necessary to separate the
transactions in the trends views by spend or income.

We couldn't just use the name of the category since they are
customizable. Typically everything except for income and transfer
will fall under spend, but this field will give us flexibility to
define them.

We also update the subcategory model test because it was
erroronously test category instead of subcategory.

We also update the db/seeds file to populate the category type. We
also remove the misc category and move the atm subcategory into
transfer since we don't want to count that as spend or income.

There are also a bunch of annotation updates that remove a newline.