Refactor Categories fixtures
This refactors the categories fixtures to enhance the names. We
also need to refactor the subcategories to reference the correct
categories. We also clean up the transactions to only keep 5 with 1
uncategorized.

This required refactoring a lot of references to the categories as
well as refactoring the transaction data entity since we removed
some transactions. No functionality in the app was changed.