Create TransactionTrendOverTime Entity
This creates a new data entity for the transaction trend over time
data. This entity is responsible for returning time series data for
transactions where the x axis is always a date.

There are four different grouping options: day, week, month and
year. Additionally, this can return spend or income data types
which is selected based on the category type.

As of now, we don't have any restrictions on the lookback time or
the number of data points that will get returned. I might consider
adding this in the future if load gets too high.

Tests were added for this entity. There is also a small update to
the category fixtures to change uncategorized to 'spend'.

We also update the Gemfile to add the groupdate gem which is
required in this entity. It's what gives us access to the
group_by_day/week/month/year methods.