Add new trends controller
This adds a new controller for returning trends data. The only
endpoint we have implemented so far is over_time, which returns
data in a time series. The endpoint sends the request straight
through to the transaction trend over time entity. The only
validation it will perform is on the date params to ensure
start_date and end_date are present.

This also makes a small update to the routes file to move the
supported_accounts endpoint as a collection under the accounts
resource.