Add user.name to account serializer
This commit refactors the account serialized object to show the
user.id and user.name in the response. This will make it easier on
the front end to display which user is the custodian of the
account. This is used a lot throughout the app since two custodians
can have the same account. Ex: both users have Chase Freedom
accounts.

We also take this opportunity to rename the .get_data method in the
account entity to .data since this is a rubocop offense.

We also make a slight update to the name of the
CategorizationRuleFilters to export the correct name of the
component.