Add react-select and standard account filter
This commit adds a new standard account filter that should be used
across the entire app. This filter takes in an array of accounts,
and will display a dropdown that allows the user to filter for
accounts. When an account is chosen, the search params are updated
with the array of account IDs to filter for. If no accounts are
selected, that indicates we aren't filtering and all accounts are
selected which removes the accounts[] query param from the
searchParams.

This pattern of using the search params for filtering specific
resources is already used across the frontend app. We just need to
start replacing the old style of filter selector for this new
style.