Add standard Subcategory filter component
This commit adds a new standard subcategory filter that should be
used across the entire app. This filter takes in an array of
categories and will display a dropdown that shows the subcategories
grouped by category. It allows the user to filter for subcategories.

When a subcategory is chosen, the search params are updated
with the array of subcategory IDs to filter for. If no subcategories
are selected, that indicates we aren't filtering and all
subcategories are selected which removes the subcategories[] query
param from the searchParams.

This pattern of using the search params for filtering specific
resources is already used across the frontend app. We just need to
start replacing the old style of filter selector for this new
style.