Add Trends page layout and filters
This commit implements the trends page layout and the filters. The
trends page will give the option to view transaction trends over
time and by category. The types of transactions are spend and
income. The left bar will be a menu that allows selecting this
trend type.

There are also 2 filters: Account and Subcategory. There is one
selector for the time range. As of now we have predefined time
ranges, but we'll explore adding custom time ranges in the future.

The graph hasn't been implemented yet and there is only a place
holder. The graph will be implemented shortly.