Add OverTimeChart to trends
This commit adds the trends Over Time chart. This trend chart will
always display time series data on the x axis. As of now, the
account or subcategory filters do not function, but are in
position. A followup commit will add logic for them to work.

We also add logic to the trends page to calculate the start and end
date based on the time range selected.