Add CategorizeTransactionsJob
This adds a new CategorizeTransactionsJob which is responsible for
loading all of the categorization rules, and iterating through each
of the uncategorized transactions.

I opted to duplicate the logic to check if a transaction matches a
rule in the job instead of the model because we avoid loading
categorization rules on every transaction. The job loads all of the
categorization rules into memory, and uses them for each
transaction. There will be magnitudes less rules than transactions.

We also install activeJob::Status to be able to monitor the status
of the job. This gem requires a store and I opted to use redis.

Redis is installed, and we configure the development and test
environments to use the localhost redis server. Production will be
configured via an env variable.

We also fix a rubocop offense in the cors.rb config file.