Add categorization jobs controller
This adds a new categorization jobs controller with two endpoints:
create and status. The create endpoint is responsible for queueing
a new categorize transactions job and returning the job ID. The
status endpoint allows fetching the status of the job and the
progress.

The ActiveJob::Status.get method will always return an object, even
if the job doesn't exist. For this, we create a fetch_job_status
helper metehod that checks if the status key is present in the
object, which indicates an existing job. If the status key is
missing, it returns nil.