Refactor categorization rule page layout
In this commit, we refactor the layout of the categorization rule
page. The rules are now the main content which takes of the left
and middle of the page. The small right column is reserved for
applying rules and the job status.

We replace the custom filters with the new react-select filters and
add labels around them. They are also moved above the rules under
the search bar.

The rules are now in a scrollable container so that it's easier to
navigate the page.