Refactor CategorizationRule components
This refactors the categorization rule components to change a couple
of things.

First, we replace the custom subcategory selector in the
EditableCategorizationRule with the new react-select version of the
SubcategorySelector.

Second, we move the subcategory section of the rules to the top of
the component instead of the bottom. Now the rule UI reads: Assign
subcategory Gym when transaction...

There are also some style changes here where we reduce padding in
the rule and increase margins between rules.