Add new SubcategorySelector component
This adds a new SubcategorySelector that is based off of
react-select rather than a custom built selector. All instances of
a subcategory selector throughout the webapp will be replaced with
this version as it's much cleaner. The props take in and return
the {id, name} of the subcategory so that there is no need to do
lookups.