Add UserFilter and refactor existing filters
This commit adds a new UserFilter. The props are an array of Users,
and as users are selected, the URL query param for users will be
updated. This is the way all filters are used across the app.

We also refactor the existing AccountFilter and SubcategoryFilter
to generalize some of the variable names. This will make it easier
in the future to copy the components and modify them to create new
filters. Each filter is slightly different so we can't create a
generalized filter, but this will make it easier to duplicate by
having to change less variables.