Refactor layout of Transaction Page
This commit refactors the transactions page to replace the custom
built filters with the new react-select filters. The filters were
moved to sit inbetween the search box and the transaction table.

We were able to get rid of the left bar since the filters are no on
top. With this, we added padding to the table so it hit the edges
of the screen. We also set a max width of the table to 6xl.

The page now shows the transaction count and filtered count from
the TransactionMetaData.

We also remove the isLoading state variables because we weren't
really implementing most of them. If we need them for real, we'll
add them back.