Refactor transaction table rows
This refactors the transaction table rows to start using the new
SubcategorySelector component instead of the home built one. Since
this was the last component using this, we can delete the
CategoryDropdown.

We also make some styling changes to the columns to minimize the
differences between the row and the expanded row.