Add subcategory filter to transaction data entity
Adds a subcategories parameter to the transaction data entity. This
will be an array of subcategory ids. When it's present, it will
only return transactions that match these subcategory ids.

We also split the spec tests for this entity into different files
because it was getting large.

Backfills more spec tests to ensure querying for the subcategory
works correctly.