Add user to expanded transaction row
This commit refactors the expanded transaction row to utilize the
formatAccountLabel helper to display the account in a uniform way.
We also add a field to this component to display the account
custodian.

In order to do this, we needed to pass the accounts array to the
expanded row component. This component will lookup the account by
the transaction.account.id.

We also needed to update the formatAccountLabel helper method to
take in an optional argument to not include the custodian's name in
the account label.