Add split columns to transaction
This commit adds two new columns to the transaction model as a
first step towards implementing the transaction split feature.

The first column, split_from_id, is a reference to the id of the
parent transaction which the split was created from.

The has_splits is a boolean flag that indicates whether or not the
transaction is a parent transaction with splits. This flag is
automatically kept up to date from Rails model callbacks
after_create and after_destroy.

We also add tests to ensure the has_splits value is kept in sync
correctly in three different cases.

This also removes the subcategory check in the
sync_category_with_subcategory callback since subcategory won't be
nil.