Add CreateSplit service
This adds a new Create Split service which is responsible for
creating transaction splits. The new transactions will be attached
to the original transactions through the .splits association. This
also sets the has_splits flag on the original transaction upon
successful split.

The service validates the split params to ensure all splits have
the same amount sign (postive/negative) as the original
transaction. We also ensure that the absolute value of the sum of
the split amounts does not exceed the absolute value of the
original transaction amount.

This service is not yet used, but will be used in an upcoming
Transactions Split Controller.