Rewrite create_split service to sync_splits service
This commit changes the way the unused splits service behaves.
Instead of only creating splits, the service now syncs splits which
behaves like an HTTP patch. It will take in an array of split_data,
and sync the original_transaction to that param data. So, if a
split transaction is specified with an id, it will attempt to
update that split. If a transaction split is passed in without an
id, it will create that split. And if a split exists on the
original transaction but isn't included in the new split_data, it
will delete that split transaction.

We also add a new exception for SplitNotFoundError which happens
when the user passes in an id of a split but that split id doesn't
exist on the original transaction.

We also update the transaction model to check for the
parent_transaction presence because when bulk destroying
transactions such as account.transactions.destroy_all, the parent
and splits will get destroyed, but then the after_destroy callbacks
will run to update the parent transactions has_splits flag, but the
parent_transaction is now nil since it was destroyed before the
callback was executed.

Fixtures for split transactions and original_transactions were
added to make testing easier.

Small fix on the category model to remove a rubocop offense.