Add transactions splits controller and sync method
This adds a new splits controller and the syncSplits endpoint. This
endpoint will be used to create, update or destroy transaction
splits. The action is a PATCH, and all splits must be provided on
every request so that we can sync the splits to the original
transaction. Existing splits won't be destroyed and recreated.
They will be updated, which is why we went with PATCH.

The endpoint handles every type of exception that the Sync Splits
service can returns, and renders the appropriate reponse and error
message.

The SplitNotFoundError will render an error message that matches
the web apps other error message format which is a response object
with an error attribute, which contains an array of field, message
objects for each error.