Walk back transaction has_splits callbacks
This commit undoes the transaction model callbacks for after_create
and after_destroy to set the has_splits flag on the parent
transaction if it exists. I'm removing this because the split
creation and deletion should only be handled through the
sync_splits service which sets the flag already. Having the call
backs in the model were causing an extra database call per split
creation.