Move all backend files up one directory
We move the Rails backend files up one directory since this
repository is only hosting the Rails backend app. We have completed
the split of the frontend, nginx, and deploy services into their
own repositories, so there is no longer a need to specify the
service name at the root level.