Add build and push script
This adds a new script to build and push the backend container
image up to our registry. The production web server will no longer
build the application image, and will now only pull it down from
the registry.

We also update the .dockerignore file to prevent copying
unnecessary files to the image for running it in production.