Transaction splits: Handle empty splits array
This adds the ability to clear a transactions splits by allowing an
empty array to be passed in as the splits parameter. We also add
logic to ensure that the splits parameter is an array, otherwise we
return a BadRequest error.

We add a test for these two edge cases.