Update image user and include db:seed in entrypoint
This updates the Dockerfile to switch the rails user to
webuser:webgroup which matches the webserver user configuration.
It's important to run these containers as non-root users for
security. The webserver will be setup with a user and group that
matches these IDs.

We also update the entrypoint script to run db:seed on every
execution to ensure the database has the necessary records for
using the app.