Validation for zero amount splits
This adds a new validation to the sync splits service to prevent
zero amount splits. This would be a pointless split and we
shouldn't allow a split to have an amount equal to $0.00. A new
sync_splits service exception class is created for this particular
validation failure.

The splits controller is also updated to handle this new exception
class.

We add a test for this in the sync_splits service test, as well as
in the splits_controller tests.